ml-service/iqa/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.0
torchvision==0.15.1

pyyaml
marshmallow-dataclass
grpcio-tools==1.53.0
click
gdown
typeguard